move all calls to start/stop metrics into the processors themselves
