+++ src/pipecat/services/ai_services.py
@@ -491,17 +491,6 @@ class ImageGenService(AIService):
     async def run_image_gen(self, prompt: str) -> AsyncGenerator[Frame, None]:
         pass
 
-    async def process_frame(self, frame: Frame, direction: FrameDirection):
-        await super().process_frame(frame, direction)
-
-        if isinstance(frame, TextFrame):
-            await self.push_frame(frame, direction)
-            await self.start_processing_metrics()
-            await self.process_generator(self.run_image_gen(frame.text))
-            await self.stop_processing_metrics(self._model if hasattr(self, "_model") else None)
-        else:
-            await self.push_frame(frame, direction)
-
 
 class VisionService(AIService):
     """VisionService is a base class for vision services."""
@@ -513,13 +502,3 @@ class VisionService(AIService):
     @abstractmethod
     async def run_vision(self, frame: VisionImageRawFrame) -> AsyncGenerator[Frame, None]:
         pass
-
-    async def process_frame(self, frame: Frame, direction: FrameDirection):
-        await super().process_frame(frame, direction)
-
-        if isinstance(frame, VisionImageRawFrame):
-            await self.start_processing_metrics()
-            await self.process_generator(self.run_vision(frame))
-            await self.stop_processing_metrics(self._model if hasattr(self, "_model") else None)
-        else:
-            await self.push_frame(frame, direction)
