+++ src/pipecat/services/lmnt.py
@@ -165,3 +165,10 @@ class LmntTTSService(AsyncTTSService):
             yield None
         except Exception as e:
             logger.exception(f"{self} exception: {e}")
+
+    async def process_frame(self, frame: Frame, direction: FrameDirection):
+        await super().process_frame(frame, direction)
+        if isinstance(frame, StartInterruptionFrame):
+            await self.stop_all_metrics()
+        else:
+            await self.push_frame(frame, direction)
