+++ src/pipecat/services/openai.py
@@ -214,6 +214,8 @@ class BaseOpenAILLMService(LLMService):
         elif isinstance(frame, LLMModelUpdateFrame):
             logger.debug(f"Switching LLM model to: [{frame.model}]")
             self._model = frame.model
+        elif isinstance(frame, StartInterruptionFrame):
+            await self.stop_all_metrics(model=self._model)
         else:
             await self.push_frame(frame, direction)
 
@@ -292,6 +294,19 @@ class OpenAIImageGenService(ImageGenService):
             frame = URLImageRawFrame(image_url, image.tobytes(), image.size, image.format)
             yield frame
 
+    async def process_frame(self, frame: Frame, direction: FrameDirection):
+        await super().process_frame(frame, direction)
+
+        if isinstance(frame, TextFrame):
+            await self.push_frame(frame, direction)
+            await self.start_processing_metrics()
+            await self.process_generator(self.run_image_gen(frame.text))
+            await self.stop_processing_metrics(model=self._model)
+        elif isinstance(frame, StartInterruptionFrame):
+            await self.stop_all_metrics(model=self._model)
+        else:
+            await self.push_frame(frame, direction)
+
 
 class OpenAITTSService(TTSService):
     """This service uses the OpenAI TTS API to generate audio from text.
@@ -354,6 +369,13 @@ class OpenAITTSService(TTSService):
                 await self.push_frame(TTSStoppedFrame())
         except BadRequestError as e:
             logger.exception(f"{self} error generating TTS: {e}")
+
+    async def process_frame(self, frame: Frame, direction: FrameDirection):
+        await super().process_frame(frame, direction)
+        if isinstance(frame, StartInterruptionFrame):
+            await self.stop_all_metrics(model=self._model)
+        else:
+            await self.push_frame(frame, direction)
 
 
 class OpenAIUserContextAggregator(LLMUserContextAggregator):
