+++ src/pipecat/services/playht.py
@@ -9,8 +9,9 @@ import struct
 
 from typing import AsyncGenerator
 
-from pipecat.frames.frames import AudioRawFrame, Frame, TTSStartedFrame, TTSStoppedFrame
+from pipecat.frames.frames import AudioRawFrame, Frame, StartInterruptionFrame, TTSStartedFrame, TTSStoppedFrame
 from pipecat.metrics.metrics import TTSUsageMetricsData
+from pipecat.processors.frame_processor import FrameDirection
 from pipecat.services.ai_services import TTSService
 
 from loguru import logger
@@ -86,3 +87,10 @@ class PlayHTTTSService(TTSService):
             await self.push_frame(TTSStoppedFrame())
         except Exception as e:
             logger.exception(f"{self} error generating TTS: {e}")
+
+    async def process_frame(self, frame: Frame, direction: FrameDirection):
+        await super().process_frame(frame, direction)
+        if isinstance(frame, StartInterruptionFrame):
+            await self.stop_all_metrics()
+        else:
+            await self.push_frame(frame, direction)
